+++ src/main/java/com/triprint/backend/domain/auth/security/controller/AuthController.java
@@ -1,0 +1,22 @@
+package com.triprint.backend.domain.auth.security.controller;
+
+import com.triprint.backend.domain.auth.security.common.ApiResponse;
+import com.triprint.backend.domain.auth.security.service.AuthService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.*;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+@RestController
+@RequestMapping("/auth")
+@RequiredArgsConstructor
+public class AuthController {
+    private final AuthService tokenService;
+
+    @GetMapping("/refresh")
+    public ApiResponse refreshToken (HttpServletRequest request, HttpServletResponse response){
+        return tokenService.refresh(request, response);
+    }
+    
+}
