+++ src/main/java/com/triprint/backend/domain/auth/security/service/AuthService.java
@@ -1,9 +1,9 @@
-package com.triprint.backend.domain.auth.security.controller;
+package com.triprint.backend.domain.auth.security.service;
 
 import com.triprint.backend.domain.auth.security.AuthToken;
 import com.triprint.backend.domain.auth.security.TokenProvider;
 import com.triprint.backend.domain.auth.security.common.ApiResponse;
-import com.triprint.backend.domain.auth.security.config.AppProperties;
+import com.triprint.backend.core.config.AppProperties;
 import com.triprint.backend.domain.user.entity.UserRefreshToken;
 import com.triprint.backend.domain.user.repository.UserRefreshTokenRepository;
 import com.triprint.backend.domain.user.status.UserRole;
@@ -11,33 +11,27 @@ import com.triprint.backend.domain.user.util.CookieUtils;
 import com.triprint.backend.domain.user.util.HeaderUtil;
 import io.jsonwebtoken.Claims;
 import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.stereotype.Service;
 
 import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.util.Date;
 
-@RestController
-@RequestMapping("/auth")
+@Service
 @RequiredArgsConstructor
-public class TokenController {
+public class AuthService {
 
     private final TokenProvider tokenProvider;
     private final AppProperties appProperties;
     private final UserRefreshTokenRepository userRefreshTokenRepository;
 
-
-    @GetMapping("/refresh")
-    public ApiResponse refreshToken (HttpServletRequest request, HttpServletResponse response){
+    public ApiResponse refresh(HttpServletRequest request, HttpServletResponse response) {
         String accessToken = HeaderUtil.getAccessToken(request);
         AuthToken authToken = tokenProvider.convertAuthToken(accessToken);
-//        if (!authToken.validate()){
-//            return ApiResponse.invalidAccessToken();
-//        }
 
         Claims claims = authToken.getExpiredTokenClaims();
-        if (claims == null){
+        if (claims == null) {
             return ApiResponse.notExpiredTokenYet();
         }
 
@@ -49,13 +43,13 @@ public class TokenController {
                 .orElse((null));
         AuthToken authRefreshToken = tokenProvider.convertRefreshToken(refreshToken);
 
-        if(!authRefreshToken.validate()){
+        if (!authRefreshToken.validate()) {
             return ApiResponse.invalidRefreshToken();
         }
 
         UserRefreshToken userRefreshToken = userRefreshTokenRepository.findByUserIdAndRefreshToken(userId, refreshToken);
-        if (userRefreshToken == null){
-            return  ApiResponse.invalidRefreshToken();
+        if (userRefreshToken == null) {
+            return ApiResponse.invalidRefreshToken();
         }
 
         Date now = new Date();
@@ -67,19 +61,28 @@ public class TokenController {
 
         long validTime = authRefreshToken.getTokenClaims().getExpiration().getTime() - now.getTime();
 
-        if (validTime <= 259200000 ){
+        if (validTime <= 259200000) {
             long refreshTokenExpiry = appProperties.getAuth().getRefreshTokenExpiration();
             authRefreshToken = tokenProvider.createAuthToken(
                     Long.parseLong(userId),
                     new Date(now.getTime() + refreshTokenExpiry)
             );
             userRefreshToken.setRefreshToken(authRefreshToken.getToken());
-            userRefreshTokenRepository.saveAndFlush(userRefreshToken);
+            userRefreshTokenRepository.saveAndFlush(userRefreshToken); //DB에 update쿼리 날리는 게 더 좋아보임 => 찾아보기
 
             int cookieMaxAge = (int) refreshTokenExpiry / 1000;
             CookieUtils.deleteCookie(request, response, "refresh_token");
             CookieUtils.addCookie(response, "refresh_token", authRefreshToken.getToken(), cookieMaxAge);
         }
-        return  ApiResponse.success("token", newAccessToken.getToken());
+
+        return ApiResponse.success("token", newAccessToken.getToken());
     }
 }
+
+
+
+
+
+
+
+
