+++ src/main/java/com/triprint/backend/domain/user/repository/UserRefreshTokenRepository.java
@@ -4,8 +4,11 @@ import com.triprint.backend.domain.user.entity.UserRefreshToken;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
 public interface UserRefreshTokenRepository extends JpaRepository<UserRefreshToken, Long> {
+//    Optional<UserRefreshToken> findById(String userId);
     UserRefreshToken findByUserId(String userId);
     UserRefreshToken findByUserIdAndRefreshToken(String userId, String refreshToken);
 }
